Use `lazy-loader`
