+++ exspy/__init__.pyi
@@ -16,29 +16,24 @@
 # You should have received a copy of the GNU General Public License
 # along with eXSpy. If not, see <https://www.gnu.org/licenses/#GPL>.
 
-"""
-Attributes
-----------
-__version__ : str
-    The version of the eXSpy package.
-preferences :
-    The global preferences for eXSpy, where the default settings are stored.
+from . import (
+    components,
+    data,
+    models,
+    signals,
+    utils,
+)
+from ._version import __version__
+from ._defaults_parser import preferences
+from ._misc import material
 
-Examples
---------
-
-To set a preference value, for example setting the EDS detector elevation
-angle to 15 degrees:
-
->>> exspy.preferences.EDS.eds_detector_elevation = 15.0
-
-To open the preferences dialog. Requires GUI widgets, to enable them see the hyperspy
-documentation for :external+hyperspy:ref:`configuring HyperSpy <configuring-hyperspy-label>`.
-
->>> exspy.preferences.gui()
-
-"""
-
-import lazy_loader
-
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
+__all__ = [
+    "__version__",
+    "components",
+    "data",
+    "preferences",
+    "material",
+    "models",
+    "signals",
+    "utils",
+]
