+++ exspy/data/__init__.py
@@ -25,6 +25,7 @@ import numpy as np
 import scipy
 
 import exspy
+from exspy import signals
 
 
 __all__ = [
@@ -168,9 +169,7 @@ def EELS_low_loss(add_noise=True, random_state=None, navigation_shape=(10,)):
     if add_noise:
         data = data + random_state.uniform(size=len(x))
 
-    from exspy.signals import EELSSpectrum
-
-    s = EELSSpectrum(data)
+    s = signals.EELSSpectrum(data)
     s.axes_manager[-1].offset = x[0]
     s.axes_manager[-1].scale = x[1] - x[0]
     s.metadata.General.title = "Artifical low loss EEL spectrum"
