+++ exspy/models/eelsmodel.py
@@ -29,7 +29,7 @@ from hyperspy.models.model1d import Model1D
 
 from exspy._docstrings.model import EELSMODEL_PARAMETERS
 from exspy.components import EELSCLEdge
-from exspy.signals.eels import EELSSpectrum
+from exspy import signals
 
 
 _logger = logging.getLogger(__name__)
@@ -116,7 +116,7 @@ class EELSModel(Model1D):
 
     @signal.setter
     def signal(self, value):
-        if isinstance(value, EELSSpectrum):
+        if isinstance(value, signals.EELSSpectrum):
             self._signal = value
         else:
             raise ValueError(
