+++ exspy/tests/test_deprecated_import.py
@@ -60,3 +60,33 @@ def test_import_eels_electron_inelastic_mean_free_path():
         )
 
     assert "use `exspy.utils.eels` instead" in record[0].message.args[0]
+
+
+def test_import_privatised_signals():
+    with pytest.warns(VisibleDeprecationWarning) as record:
+        from exspy.signals.eels import EELSSpectrum, LazyEELSSpectrum
+
+    assert "use `exspy.signals` instead" in record[0].message.args[0]
+
+    with pytest.warns(VisibleDeprecationWarning) as record:
+        from exspy.signals.dielectric_function import (
+            DielectricFunction,
+            LazyDielectricFunction,
+        )
+
+    assert "use `exspy.signals` instead" in record[0].message.args[0]
+
+    with pytest.warns(VisibleDeprecationWarning) as record:
+        from exspy.signals.eds import EDSSpectrum, LazyEDSSpectrum
+
+    assert "use `exspy.signals` instead" in record[0].message.args[0]
+
+    with pytest.warns(VisibleDeprecationWarning) as record:
+        from exspy.signals.eds_tem import EDSTEMSpectrum, LazyEDSTEMSpectrum
+
+    assert "use `exspy.signals` instead" in record[0].message.args[0]
+
+    with pytest.warns(VisibleDeprecationWarning) as record:
+        from exspy.signals.eds_sem import EDSSEMSpectrum, LazyEDSSEMSpectrum
+
+    assert "use `exspy.signals` instead" in record[0].message.args[0]
